pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-function-msgraphhelper"
version = "0.0.1"
description = "Azure Function MS Graph Helper"
authors = [{ name = "Giles Antonio Radford", email = "moof@metamoof.net" }]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["azure-functions", "ms-graph", "graph", "suscription"]
dependencies = ["azure-functions", "azure-identity", "azure-data-tables"]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = ["pytest", "black"]
